+++ app/routes/sketch+/artboards+/$slug_+/actions/artboard-design-template.ts
@@ -1,112 +1,12 @@
 import { json } from '@remix-run/node'
 import { type IntentActionArgs } from '#app/definitions/intent-action-args'
-import {
-	ArtboardSelectedDesignsSchema,
-	type ArtboardSelectedDesignsType,
-} from '#app/schema/artboard'
-import { NewArtboardDesignSchema, designSchema } from '#app/schema/design'
 import { EditArtboardTemplateStyleSchema } from '#app/schema/template'
 import {
 	notSubmissionResponse,
 	submissionErrorResponse,
 } from '#app/utils/conform-utils'
-import { prisma } from '#app/utils/db.server'
 import { findFirstTemplateInstance } from '#app/utils/prisma-extensions-template'
 import { parseArtboardDesignSubmission } from './utils'
-
-export async function artboardDesignNewTemplateAction({
-	userId,
-	formData,
-}: IntentActionArgs) {
-	// validation
-	const submission = await parseArtboardDesignSubmission({
-		userId,
-		formData,
-		schema: NewArtboardDesignSchema,
-	})
-	if (submission.intent !== 'submit') {
-		return notSubmissionResponse(submission)
-	}
-	if (!submission.value) {
-		return submissionErrorResponse(submission)
-	}
-
-	// changes
-	const { artboardId } = submission.value
-
-	// start transaction so we can create design and template together
-	// template is 1:1 with design which belongs to an artboard
-	try {
-		await prisma.$transaction(async prisma => {
-			// new designs are appended to the end of the list
-			// find the last design in the list by type
-			// we know the artboard already exists for the user by this point
-			const lastArtboardDesignTemplate = await prisma.design.findFirst({
-				where: { type: 'template', artboardId, nextId: null },
-			})
-
-			// create design before template
-			const designData = designSchema.parse({
-				type: 'template',
-				ownerId: userId,
-				artboardId,
-			})
-			const design = await prisma.design.create({
-				data: designData,
-			})
-
-			// then create template after design
-			await prisma.template.create({
-				data: {
-					designId: design.id,
-				},
-			})
-
-			// if the artboard already has a template
-			// link the new template to the last one
-			// and the last one to the new one
-			if (lastArtboardDesignTemplate) {
-				await prisma.design.update({
-					where: { id: design.id },
-					data: { prevId: lastArtboardDesignTemplate.id },
-				})
-
-				await prisma.design.update({
-					where: { id: lastArtboardDesignTemplate.id },
-					data: { nextId: design.id },
-				})
-			} else {
-				// set selectedDesign for the artboard
-				// get the selectedDesigns from the artboard
-				const artboard = await prisma.artboard.findFirst({
-					where: { id: artboardId },
-					select: { selectedDesigns: true },
-				})
-				if (!artboard) return submissionErrorResponse(submission)
-
-				// parse the selectedDesigns and update the paletteId
-				const selectedDesigns = ArtboardSelectedDesignsSchema.parse(
-					JSON.parse(artboard.selectedDesigns),
-				) as ArtboardSelectedDesignsType
-				const updatedSelectedDesigns = ArtboardSelectedDesignsSchema.parse({
-					...selectedDesigns,
-					templateId: design.id,
-				})
-
-				// update the selectedDesigns for the artboard
-				await prisma.artboard.update({
-					where: { id: artboardId },
-					data: { selectedDesigns: JSON.stringify(updatedSelectedDesigns) },
-				})
-			}
-		})
-	} catch (error) {
-		console.log(error)
-		return submissionErrorResponse(submission)
-	}
-
-	return json({ status: 'success', submission } as const)
-}
 
 export async function artboardDesignEditTemplateStyleAction({
 	userId,
