+++ app/routes/sketch+/artboards+/$slug_+/actions/artboard-design.ts
@@ -4,8 +4,8 @@ import {
 	DeleteArtboardDesignSchema,
 	NewArtboardDesignSchema,
 	ReorderArtboardDesignSchema,
-	ToggleVisibilityArtboardDesignSchema,
-} from '#app/schema/design'
+	ToggleVisibleArtboardDesignSchema,
+} from '#app/schema/design-artboard'
 import {
 	notSubmissionResponse,
 	submissionErrorResponse,
@@ -30,7 +30,7 @@ async function validateSubmission({
 	schema:
 		| typeof NewArtboardDesignSchema
 		| typeof ReorderArtboardDesignSchema
-		| typeof ToggleVisibilityArtboardDesignSchema
+		| typeof ToggleVisibleArtboardDesignSchema
 		| typeof DeleteArtboardDesignSchema
 }) {
 	const newDesign = schema === NewArtboardDesignSchema
@@ -107,7 +107,7 @@ export async function artboardDesignToggleVisibilityAction({
 	const { submission, isValid, response } = await validateSubmission({
 		userId,
 		formData,
-		schema: ToggleVisibilityArtboardDesignSchema,
+		schema: ToggleVisibleArtboardDesignSchema,
 	})
 	if (!isValid || !submission) return response
