+++ app/routes/sketch+/artboards+/$slug_+/components/forms/artboard/design/panel-form-artboard-design-reorder.tsx
@@ -5,29 +5,36 @@ import { useActionData, useFetcher } from '@remix-run/react'
 import { AuthenticityTokenInput } from 'remix-utils/csrf/react'
 import { Button } from '#app/components/ui/button'
 import { Icon } from '#app/components/ui/icon'
-import { DeleteArtboardDesignSchema } from '#app/schema/design'
+import { ReorderArtboardDesignSchema } from '#app/schema/design-artboard'
 import { useIsPending } from '#app/utils/misc'
-import { INTENT } from '../intent'
-import { type action } from '../route'
+import { ARTBOARD_DESIGN_INTENT } from '../../../../intent'
+import { type action } from '../../../../route'
 
-export const PanelFormArtboardDesignDelete = ({
+export const PanelFormArtboardDesignReorder = ({
 	id,
 	artboardId,
-	isSelectedDesign,
+	panelIndex,
+	panelCount,
+	direction,
 	updateSelectedDesignId,
 }: {
 	id: string
 	artboardId: Artboard['id']
-	isSelectedDesign: boolean
+	panelIndex: number
+	panelCount: number
+	direction: 'up' | 'down'
 	updateSelectedDesignId: string | null | undefined
 }) => {
 	const fetcher = useFetcher<typeof action>()
 	const actionData = useActionData<typeof action>()
 	const isPending = useIsPending()
 
+	const atTop = panelIndex === 0 && direction === 'up'
+	const atBottom = panelIndex === panelCount - 1 && direction === 'down'
+
 	const [form] = useForm({
-		id: `panel-form-artboard-design-delete-${id}`,
-		constraint: getFieldsetConstraint(DeleteArtboardDesignSchema),
+		id: `panel-form-artboard-design-reorder-${id}-${direction}`,
+		constraint: getFieldsetConstraint(ReorderArtboardDesignSchema),
 		lastSubmission: actionData?.submission,
 	})
 
@@ -37,13 +44,7 @@ export const PanelFormArtboardDesignDelete = ({
 
 			<input type="hidden" name="id" value={id} />
 			<input type="hidden" name="artboardId" value={artboardId} />
-			{isSelectedDesign && (
-				<input
-					type="hidden"
-					name="isSelectedDesign"
-					value={String(isSelectedDesign)}
-				/>
-			)}
+			<input type="hidden" name="direction" value={direction} />
 			{updateSelectedDesignId && (
 				<input
 					type="hidden"
@@ -51,15 +52,19 @@ export const PanelFormArtboardDesignDelete = ({
 					value={updateSelectedDesignId}
 				/>
 			)}
-			<input type="hidden" name="intent" value={INTENT.artboardDeleteDesign} />
+			<input
+				type="hidden"
+				name="intent"
+				value={ARTBOARD_DESIGN_INTENT.artboardReorderDesign}
+			/>
 			<Button
 				type="submit"
 				variant="ghost"
-				className="flex h-8 w-8 cursor-pointer items-center justify-center"
-				disabled={isPending}
+				className="flex h-4 w-4 cursor-pointer items-center justify-center"
+				disabled={isPending || atTop || atBottom}
 			>
-				<Icon name="minus">
-					<span className="sr-only">Remove</span>
+				<Icon name={`chevron-${direction}`}>
+					<span className="sr-only">Move {direction}</span>
 				</Icon>
 			</Button>
 		</fetcher.Form>
