+++ app/routes/sketch+/artboards+/$slug_+/components/panel/artboard/design/panel-content-artboard-design-template.tsx
@@ -9,27 +9,26 @@ import {
 	PanelTitle,
 } from '#app/components/shared'
 import { Input } from '#app/components/ui/input'
-import { type IDesignWithRotate } from '#app/models/design.server'
+import { type IDesignWithTemplate } from '#app/models/design.server'
 import { DesignTypeEnum } from '#app/schema/design'
 import {
 	panelItemVariablesDesignType,
 	panelListVariablesDesignType,
 	selectedDesignsOnUpdate,
 } from '#app/utils/design'
-import { type PickedArtboardType } from '../queries'
-import { PanelFormArtboardDesignDelete } from './panel-form-artboard-design-delete'
-import { PanelFormArtboardDesignEditRotate } from './panel-form-artboard-design-edit-rotate'
-import { PanelFormArtboardDesignNew } from './panel-form-artboard-design-new'
-import { PanelFormArtboardDesignReorder } from './panel-form-artboard-design-reorder'
-import { PanelFormArtboardDesignToggleVisibility } from './panel-form-artboard-design-toggle-visibility'
-import { PanelPopoverArtboardDesignRotate } from './panel-popover-artboard-design-rotate'
+import { type PickedArtboardType } from '../../../../queries'
+import { PanelFormArtboardDesignDelete } from '../../../forms/artboard/design/panel-form-artboard-design-delete'
+import { PanelFormArtboardDesignNew } from '../../../forms/artboard/design/panel-form-artboard-design-new'
+import { PanelFormArtboardDesignReorder } from '../../../forms/artboard/design/panel-form-artboard-design-reorder'
+import { PanelFormArtboardDesignToggleVisible } from '../../../forms/artboard/design/panel-form-artboard-design-toggle-visible'
+import { PanelPopoverDesignTemplate } from '../../../popovers/design/panel-popover-design-template'
 
-export const PanelContentArtboardDesignRotate = ({
+export const PanelContentArtboardDesignTemplate = ({
 	artboard,
-	designRotates,
+	designTemplates,
 }: {
 	artboard: PickedArtboardType
-	designRotates: IDesignWithRotate[]
+	designTemplates: IDesignWithTemplate[]
 }) => {
 	const {
 		orderedDesigns,
@@ -39,25 +38,25 @@ export const PanelContentArtboardDesignRotate = ({
 		firstVisibleDesignId,
 		selectedDesignId,
 	} = panelListVariablesDesignType({
-		designs: designRotates,
+		designs: designTemplates,
 		artboard,
-		type: DesignTypeEnum.ROTATE,
+		type: DesignTypeEnum.TEMPLATE,
 	})
 
 	return (
 		<Panel>
 			<PanelHeader>
-				<PanelTitle>Rotate</PanelTitle>
+				<PanelTitle>Template</PanelTitle>
 				<div className="flex flex-shrink">
 					<PanelFormArtboardDesignNew
 						artboardId={artboard.id}
-						type={DesignTypeEnum.ROTATE}
+						type={DesignTypeEnum.TEMPLATE}
 						visibleDesignsCount={visibleDesignIds.length}
 					/>
 				</div>
 			</PanelHeader>
 			{orderedDesigns.map((design, index) => {
-				const { id, visible, rotate } = design as IDesignWithRotate
+				const { id, visible, template } = design as IDesignWithTemplate
 
 				const {
 					isSelectedDesign,
@@ -89,7 +88,7 @@ export const PanelContentArtboardDesignRotate = ({
 				})
 
 				return (
-					<PanelRow key={rotate.id}>
+					<PanelRow key={template.id}>
 						<PanelRowOrderContainer>
 							<PanelFormArtboardDesignReorder
 								id={id}
@@ -110,26 +109,16 @@ export const PanelContentArtboardDesignRotate = ({
 						</PanelRowOrderContainer>
 						<PanelRowContainer>
 							<PanelRowValueContainer>
-								<PanelPopoverArtboardDesignRotate
-									artboardId={artboard.id}
-									rotate={rotate}
+								<PanelPopoverDesignTemplate template={template} />
+								<Input
+									type="text"
+									className={'flex h-8'}
+									disabled
+									defaultValue={template.style}
 								/>
-								{rotate.basis !== 'defined' ? (
-									<Input
-										type="text"
-										className={'flex h-8'}
-										disabled
-										defaultValue={rotate.basis}
-									/>
-								) : (
-									<PanelFormArtboardDesignEditRotate
-										artboardId={artboard.id}
-										rotate={rotate}
-									/>
-								)}
 							</PanelRowValueContainer>
 							<PanelRowIconContainer>
-								<PanelFormArtboardDesignToggleVisibility
+								<PanelFormArtboardDesignToggleVisible
 									id={id}
 									artboardId={artboard.id}
 									visible={visible}
