+++ app/routes/sketch+/artboards+/$slug_+/components/panel/artboard/panel-content-artboard-designs.tsx
@@ -2,14 +2,14 @@ import { type IDesignWithType } from '#app/models/design.server'
 import { filterAndOrderArtboardDesignsByType } from '#app/utils/design'
 import { type PickedArtboardType } from '../../../queries'
 import { PanelContentArtboardBackgroundColor } from '../../panel-content-artboard-background-color'
-import { PanelContentArtboardDesignFill } from '../../panel-content-artboard-design-fill'
-import { PanelContentArtboardDesignLayout } from '../../panel-content-artboard-design-layout'
-import { PanelContentArtboardDesignLine } from '../../panel-content-artboard-design-line'
-import { PanelContentArtboardDesignPalette } from '../../panel-content-artboard-design-palette'
-import { PanelContentArtboardDesignRotate } from '../../panel-content-artboard-design-rotate'
-import { PanelContentArtboardDesignSize } from '../../panel-content-artboard-design-size'
-import { PanelContentArtboardDesignStroke } from '../../panel-content-artboard-design-stroke'
-import { PanelContentArtboardDesignTemplate } from '../../panel-content-artboard-design-template'
+import { PanelContentArtboardDesignFill } from './design/panel-content-artboard-design-fill'
+import { PanelContentArtboardDesignLayout } from './design/panel-content-artboard-design-layout'
+import { PanelContentArtboardDesignLine } from './design/panel-content-artboard-design-line'
+import { PanelContentArtboardDesignPalette } from './design/panel-content-artboard-design-palette'
+import { PanelContentArtboardDesignRotate } from './design/panel-content-artboard-design-rotate'
+import { PanelContentArtboardDesignSize } from './design/panel-content-artboard-design-size'
+import { PanelContentArtboardDesignStroke } from './design/panel-content-artboard-design-stroke'
+import { PanelContentArtboardDesignTemplate } from './design/panel-content-artboard-design-template'
 import { PanelContentArtboardFrame } from './panel-content-artboard-frame'
 
 export const PanelContentArtboardDesigns = ({
