+++ app/schema/design.ts
@@ -38,31 +38,6 @@ export const LayerDesignDataCreateSchema = z.object({
 	selected: z.boolean().optional(),
 }) satisfies z.Schema<Design>
 
-export const NewArtboardDesignSchema = z.object({
-	artboardId: z.string(),
-	type: z.nativeEnum(DesignTypeEnum),
-	visibleDesignsCount: z.number(),
-})
-
-export const DeleteArtboardDesignSchema = z.object({
-	id: z.string(),
-	artboardId: z.string(),
-	updateSelectedDesignId: z.string().optional(),
-})
-
-export const ToggleVisibilityArtboardDesignSchema = z.object({
-	id: z.string(),
-	artboardId: z.string(),
-	updateSelectedDesignId: z.string().optional(),
-})
-
-export const ReorderArtboardDesignSchema = z.object({
-	id: z.string(),
-	artboardId: z.string(),
-	direction: z.enum(['up', 'down']),
-	updateSelectedDesignId: z.string().optional(),
-})
-
 export type selectArgsType = z.infer<typeof selectArgs>
 const selectArgs = z.object({
 	id: z.boolean().optional(),
